Fix CI with newest molecule (#303)
* Fix CI with newest molecule

Get rid of the use of  molecule ephemeral_directory() as it's gone.
The most affected test was the vagrant one, as it's checking the
generated Vagrantfile. Set the environment variable for the
ephemeral directory in this case, so that we know for sure
the path of the ephemeral directory.

Signed-off-by: Arnaud Patard <arnaud.patard@collabora.com>

* Update prettier pre-commit configuration

The repository has been archived, so don't use it.

Signed-off-by: Arnaud Patard <arnaud.patard@collabora.com>

---------

Signed-off-by: Arnaud Patard <arnaud.patard@collabora.com>