fix: openstack driver private address retrieval
In ansible 2.19, the template for retrieving the public(floating) address
of a molecule instance returns a NoneType value, which in turn evaluates
to False when compared to , which cause the private address of the instance
to never get picker up as fallback. This commit changes the condition to
not address | bool which in both case will evaluate to True if the retrieval
of the public ip failed.